(BOLT-1173) Split YamlPlanEvaluator into separate classes
This splits the YamlPlanEvaluator into three new classes:
* YamlPlan, the model class
* Loader, which parses YAML and instantiates YamlPlan objects
* Evaluator, which evaluates embedded Puppet code and runs each step
